♻️ Refactor: #123 Update Category imports to new architecture
- Update app/sitemap.ts to import from modules/post/domain
- Update app/category/[category]/page.tsx to import from modules/post/domain
- Update components/header/Header.tsx to import from modules/post/domain
